librustc: Allow linkage attribute on any statics, not just foreign statics.
